requirements.txt
connexion
pyyaml
pytest
pygame
flask_testing
editdistance
phonetics
click
